outsourced prefilter. Added deploymentconfig to it
