+++ controllers/deploymentConfig_watch_controller.go
@@ -20,6 +20,7 @@ import (
 	"github.com/containersol/prescale-operator/internal/reconciler"
 	"github.com/containersol/prescale-operator/internal/resources"
 	"github.com/containersol/prescale-operator/internal/states"
+	"github.com/containersol/prescale-operator/internal/validations"
 	"github.com/go-logr/logr"
 	ocv1 "github.com/openshift/api/apps/v1"
 
@@ -95,7 +96,7 @@ func (r *DeploymentConfigWatcher) Reconcile(ctx context.Context, req ctrl.Reques
 	}
 
 	// After we have the deploymentconfig and state data, we are ready to reconcile the deploymentconfig
-	err = reconciler.ReconcileDeploymentConfig(ctx, r.Client, deploymentconfig, finalState)
+	err = reconciler.ReconcileDeploymentConfig(ctx, r.Client, deploymentconfig, finalState, optinLabel)
 	if err != nil {
 		return ctrl.Result{}, err
 	}
@@ -109,5 +110,6 @@ func (r *DeploymentConfigWatcher) Reconcile(ctx context.Context, req ctrl.Reques
 func (r *DeploymentConfigWatcher) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&ocv1.DeploymentConfig{}).
+		WithEventFilter(validations.PreFilter()).
 		Complete(r)
 }
