+++ controllers/deployment_watch_controller.go
@@ -20,7 +20,7 @@ import (
 	"github.com/containersol/prescale-operator/internal/reconciler"
 	"github.com/containersol/prescale-operator/internal/resources"
 	"github.com/containersol/prescale-operator/internal/states"
-	"github.com/containersol/prescale-operator/pkg/utils/labels"
+	"github.com/containersol/prescale-operator/internal/validations"
 
 	"github.com/go-logr/logr"
 	v1 "k8s.io/api/apps/v1"
@@ -29,9 +29,7 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
-	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 )
 
@@ -49,47 +47,6 @@ type DeploymentWatcher struct {
 func (r *DeploymentWatcher) WatchForDeployments(client client.Client, c controller.Controller) error {
 
 	return c.Watch(&source.Kind{Type: &v1.Deployment{}}, &handler.EnqueueRequestForObject{})
-}
-
-// PreFilter for incoming changes of deployments. We only care about changes on opt-in label.
-func PreFilter() predicate.Predicate {
-	return predicate.Funcs{
-		UpdateFunc: func(e event.UpdateEvent) bool {
-			// Ignore updates to CR status in which case metadata.Generation does not change
-			deploymentName := e.ObjectNew.GetName()
-
-			var oldoptin, newoptin bool
-			newoptin = labels.GetLabelValue(e.ObjectNew.GetLabels(), "scaler/opt-in")
-			oldoptin = labels.GetLabelValue(e.ObjectOld.GetLabels(), "scaler/opt-in")
-
-			if newoptin == true {
-				log := ctrl.Log
-				log.Info("(UpdateEvent) Deployment is opted in. Reconciling.. " + deploymentName)
-				return true
-			} else if oldoptin == true && newoptin == false {
-				log := ctrl.Log.
-					WithValues("old", oldoptin).
-					WithValues("new", newoptin)
-				log.Info("(UpdateEvent) Labels are for deployment. Deployment" + deploymentName + " opted out. Trying to reconcile back to default replica count")
-				return true
-			}
-
-			return false
-		},
-		CreateFunc: func(e event.CreateEvent) bool {
-			newlabels := e.Object.GetLabels()
-
-			newoptin := labels.GetLabelValue(newlabels, "scaler/opt-in")
-			deploymentName := e.Object.GetName()
-			log := ctrl.Log
-			log.Info("(CreateEvent) New opted-in deployment detected. Reconciling.. " + deploymentName)
-
-			return newoptin
-		},
-		DeleteFunc: func(e event.DeleteEvent) bool {
-			return false
-		},
-	}
 }
 
 // Reconcile tries to reconcile the replicas of the opted-in deployments
@@ -147,6 +104,6 @@ func (r *DeploymentWatcher) Reconcile(ctx context.Context, req ctrl.Request) (ct
 func (r *DeploymentWatcher) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&v1.Deployment{}).
-		WithEventFilter(PreFilter()).
+		WithEventFilter(validations.PreFilter()).
 		Complete(r)
 }
